1.修改数据

+++ app/src/main/java/cn/xxjc/com/bean/TableCols.java
@@ -12,7 +12,7 @@ public class TableCols implements Comparable {
     public int tableId;
     public int minNum;
     public int maxNum;
-    public int value=-1111;
+    public double value=-1111;
     public int tableColsValueId;
     public int type;
