+++ app/src/main/java/cn/xxjc/com/utils/XmlParseUtil.java
@@ -184,7 +184,7 @@ public class XmlParseUtil {
                         tableColsValue.tablename=pullParser.nextText();
                     }
                     if("colValue".equals(pullParser.getName())){
-                        tableColsValue.colValue= Integer.parseInt(pullParser.nextText());
+                        tableColsValue.colValue= Double.parseDouble(pullParser.nextText());
                     }
 
                     break;
@@ -230,7 +230,7 @@ public class XmlParseUtil {
                         tableColsValue.tablename=pullParser.nextText();
                     }
                     if("colValue".equals(pullParser.getName())){
-                        tableColsValue.colValue= Integer.parseInt(pullParser.nextText());
+                        tableColsValue.colValue= Double.parseDouble(pullParser.nextText());
                     }
 
                     break;
